Move time converter function to functions file
